refactor(component): keep audit split finding-neutral
Merge RequestedDetectorRule types into source_policy.rs so the file
retains *Config types and the naming-convention detector does not flag
it. Genericize ecosystem-term examples in detector_profile.rs and
known_symbols.rs doc comments that the monolith previously masked behind
its early #[cfg(test)] scan-stop, keeping the core-agnostic-source audit
finding count neutral.
